Update the file structure.
